OPHJOD-410: Refactor
