+++ src/main/java/fi/okm/jod/yksilo/entity/YksilonOsaaminen.java
@@ -9,10 +9,12 @@
 
 package fi.okm.jod.yksilo.entity;
 
+import fi.okm.jod.yksilo.domain.OsaamisenLahdeTyyppi;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
@@ -20,6 +22,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.NotNull;
 import java.util.UUID;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
 @Entity
@@ -28,7 +31,7 @@ public class YksilonOsaaminen {
   @Id @GeneratedValue private UUID id;
 
   @Getter
-  @ManyToOne
+  @ManyToOne(fetch = FetchType.LAZY)
   @NotNull
   @JoinColumn(nullable = false)
   private Yksilo yksilo;
@@ -39,6 +42,7 @@ public class YksilonOsaaminen {
   @JoinColumn(nullable = false)
   private Osaaminen osaaminen;
 
+  @Setter
   @Enumerated(EnumType.STRING)
   @NotNull
   @Column(nullable = false)
@@ -60,10 +64,12 @@ public class YksilonOsaaminen {
   public void setKoulutus(Koulutus koulutus) {
     this.lahde = OsaamisenLahdeTyyppi.KOULUTUS;
     this.koulutus = koulutus;
+    this.toimenkuva = null;
   }
 
   public void setToimenkuva(Toimenkuva toimenkuva) {
     this.lahde = OsaamisenLahdeTyyppi.TOIMENKUVA;
     this.toimenkuva = toimenkuva;
+    this.koulutus = null;
   }
 }
