+++ src/main/java/fi/okm/jod/yksilo/validator/LocalizedStringSizeValidator.java
@@ -9,7 +9,7 @@
 
 package fi.okm.jod.yksilo.validator;
 
-import fi.okm.jod.yksilo.dto.LocalizedString;
+import fi.okm.jod.yksilo.domain.LocalizedString;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import jakarta.validation.constraints.Size;
