+++ src/main/java/fi/okm/jod/yksilo/validator/NotEmptyLocalizedStringValidator.java
@@ -9,7 +9,7 @@
 
 package fi.okm.jod.yksilo.validator;
 
-import fi.okm.jod.yksilo.dto.LocalizedString;
+import fi.okm.jod.yksilo.domain.LocalizedString;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import jakarta.validation.constraints.NotEmpty;
